lot of refactoring

+++ Calendo/GoogleCalendar/GoogleCalendar.cs
@@ -19,30 +19,33 @@ namespace Calendo.GoogleCalendar
 {
     class GoogleCalendar
     {
-        //private static string STORAGE_PATH = "archive.txt";
         private TaskManager storage = TaskManager.Instance;
-       private static string auth = "";
-       // private static String taskListId = getTaskListId(auth);
-    
-        // Reserved for duplex sync
-        public string Sync()
+        private static string auth = "";
+   
+        public Boolean Export()
         {
-            Export();
-            //Import();
-            /*return "";
-            storage.Load();
-            string auth = Authorize();
-            List<String> tasks = getTasksIds(getTaskResponse(auth));
-            postTasks(storage.Entries, auth);
-            deleteGcalTasks(tasks, auth);*/
-            return "";
-        }
-
-        public void Export()
-        {
+            if (auth == "")
+                return false;
             storage.Load();
             deleteGcalTasks( getTasksIds(getTaskResponse(auth)),auth);
             postTasks(storage.Entries, auth);
+            return true;
+        }
+
+        public Boolean Import()
+        {
+            if (auth == "")
+                return false;
+            List<Entry> taskList = getTaskDetails(getTaskResponse(auth));
+            storage.Load();
+            storage.Entries.Clear();
+
+            foreach (Entry task in taskList)
+            {
+                storage.Entries.Add(task);
+            }
+            storage.Save();
+            return true;
         }
 
         private string getTaskResponse(string auth)
@@ -71,20 +74,6 @@ namespace Calendo.GoogleCalendar
                     tasks += sLine;
             }
             return tasks;
-        }
-
-        public string Import()
-        {
-            List<Entry> taskList = getTaskDetails(getTaskResponse(auth));
-            storage.Load();
-            storage.Entries.Clear();
-
-            foreach (Entry task in taskList)
-            {
-                storage.Entries.Add(task);
-            }
-            storage.Save();
-            return "";     
         }
 
         private string getTaskListId(String auth)
@@ -131,7 +120,7 @@ namespace Calendo.GoogleCalendar
         private static string GetAuthentication(NativeApplicationClient provider)
         {
             string url = "https://accounts.google.com/o/oauth2/auth?";
-            url += "scope=https://www.googleapis.com/auth/userinfo.email+https://www.googleapis.com/auth/userinfo.profile+https://www.googleapis.com/auth/tasks";
+            url += "scope=https://www.googleapis.com/auth/userinfo.email+https://www.googleapis.com/auth/userinfo.profile+https://www.googleapis.com/auth/tasks+https://www.googleapis.com/auth/calendar";
             url += "&redirect_uri=http://rahij.com/calendo.php";
             url += "&response_type=token";
             url += "&client_id=" + provider.ClientIdentifier;
@@ -146,7 +135,6 @@ namespace Calendo.GoogleCalendar
 
             if (authCode == "")
             {
-                MessageBox.Show("Authorization code not provided!");
                 return "";
             }
             // Retrieve the access token by using the authorization code:
@@ -160,30 +148,33 @@ namespace Calendo.GoogleCalendar
             foreach (Entry task in tasks)
             {
                 String taskListId = getTaskListId(auth);
+                /*
+                 * Creating a http request
+                 */
+
                 HttpWebRequest httpWReq =
                     (HttpWebRequest)WebRequest.Create("https://www.googleapis.com/tasks/v1/lists/" + taskListId + "/tasks?key=AIzaSyDQPMYzYwXWh4JUZX16RnV2DNJddg_5INo&access_token=" + auth);
-
                 httpWReq.ContentType = "application/json";
-                ASCIIEncoding encoding = new ASCIIEncoding();
-
-                responseText = "";
-                JSON<TaskResponse> jtest = new JSON<TaskResponse>();
+                httpWReq.Method = "POST";
+                httpWReq.Timeout = 10000;
+                JSON<TaskResponse> jsonObject = new JSON<TaskResponse>();
                 string postData = "{\"title\": \"" + task.Description + "\"";
-                if(task.Type!= EntryType.FLOATING)
-                    postData+=",\"due\": \"" + jtest.DateToJSON(task.StartTime) + "\"";
-                postData+="}";
+                if (task.Type != EntryType.FLOATING)
+                    postData += ",\"due\": \"" + jsonObject.DateToJSON(task.StartTime) + "\"";
+                postData += "}";
+                responseText = "";
 
+                ASCIIEncoding encoding = new ASCIIEncoding();
                 byte[] data = encoding.GetBytes(postData);
 
-                httpWReq.Method = "POST";
                 httpWReq.ContentLength = data.Length;
-
                 using (Stream newStream = httpWReq.GetRequestStream())
                 {
                     newStream.Write(data, 0, data.Length);
                 }
 
                 var httpResponse = (HttpWebResponse)httpWReq.GetResponse();
+
                 using (var streamReader = new StreamReader(httpResponse.GetResponseStream()))
                 {
                     responseText += streamReader.ReadToEnd();
@@ -196,22 +187,20 @@ namespace Calendo.GoogleCalendar
 		private void deleteGcalTasks(List<String> taskIds, string auth)
         {
             string taskListId = getTaskListId(auth);
-            storage.Load();
             foreach (string taskId in taskIds)
             {
-                HttpWebRequest httpWReq =
-(HttpWebRequest)WebRequest.Create("https://www.googleapis.com/tasks/v1/lists/" + taskListId + "/tasks/" + taskId + "?key=AIzaSyDQPMYzYwXWh4JUZX16RnV2DNJddg_5INo&access_token=" + auth);
-
-                ASCIIEncoding encoding = new ASCIIEncoding();
+                HttpWebRequest httpWReq = (HttpWebRequest)WebRequest.Create("https://www.googleapis.com/tasks/v1/lists/" + taskListId + "/tasks/" + taskId + "?key=AIzaSyDQPMYzYwXWh4JUZX16RnV2DNJddg_5INo&access_token=" + auth);
                 httpWReq.Method = "DELETE";
+                httpWReq.ContentType = "application/x-www-form-urlencoded";
                 HttpWebResponse response = (HttpWebResponse)httpWReq.GetResponse();
+                MessageBox.Show(response.StatusDescription);
             }
         }
 		
 		private List<String> getTasksIds(string tasks)
 		{
-            JSON<TaskResponse> jtest = new JSON<TaskResponse>();
-            TaskResponse values = jtest.Deserialize(tasks);
+            JSON<TaskResponse> jsonObject = new JSON<TaskResponse>();
+            TaskResponse values = jsonObject.Deserialize(tasks);
             List<String> taskList = new List<string>();
             for (int c = 0; c < values.Items.Count; c++)
             {
@@ -244,5 +233,6 @@ namespace Calendo.GoogleCalendar
             }
             return taskList;
         }
+
     }
 }
